Module mmoshforge::accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

ACreateCollection

Generated client accounts for ACreateCollection.

ACreatePassToken

Generated client accounts for ACreatePassToken.

AInitActivationToken

Generated client accounts for AInitActivationToken.

AInitMainState

Generated client accounts for AInitMainState.

AInitPassToken

Generated client accounts for AInitPassToken.

AMintActivationToken

Generated client accounts for AMintActivationToken.

AMintGuestPass

Generated client accounts for AMintGuestPass.

AMintPassByAdmin

Generated client accounts for AMintPassByAdmin.

AMintPassByAt

Generated client accounts for AMintPassByAt.

AMintProfileByAdmin

Generated client accounts for AMintProfileByAdmin.

AMintProfileByAt

Generated client accounts for AMintProfileByAt.

AUpdateMainState

Generated client accounts for AUpdateMainState.

AUpdateMainStateOwner

Generated client accounts for AUpdateMainStateOwner.

AUpdateMint

Generated client accounts for AUpdateMint.

BuyCommonV0

Generated client accounts for BuyCommonV0.

BuyLaunchPass

Generated client accounts for BuyLaunchPass.

BuyNativeV0

Generated client accounts for BuyNativeV0.

BuyV1

Generated client accounts for BuyV1.

BuyWrappedSolV0

Generated client accounts for BuyWrappedSolV0.

CloseTokenBondingV0

Generated client accounts for CloseTokenBondingV0.

InitLaunchPass

Generated client accounts for InitLaunchPass.

InitVault

Generated client accounts for InitVault.

InitializeCurveV0

Generated client accounts for InitializeCurveV0.

InitializeSolStorageV0

Generated client accounts for InitializeSolStorageV0.

InitializeTokenBondingV0

Generated client accounts for InitializeTokenBondingV0.

RedeemLaunchPass

Generated client accounts for RedeemLaunchPass.

SellCommonV0

Generated client accounts for SellCommonV0.

SellNativeV0

Generated client accounts for SellNativeV0.

SellV1

Generated client accounts for SellV1.

SellWrappedSolV0

Generated client accounts for SellWrappedSolV0.

StakeVault

Generated client accounts for StakeVault.

TransferReservesNativeV0

Generated client accounts for TransferReservesNativeV0.

TransferReservesV0

Generated client accounts for TransferReservesV0.

TransferReservesV0Common

Generated client accounts for TransferReservesV0Common.

UnstakeVault

Generated client accounts for UnstakeVault.

UpdateCurveV0

Generated client accounts for UpdateCurveV0.

UpdateReserveAuthorityV0

Generated client accounts for UpdateReserveAuthorityV0.

UpdateTokenBondingV0

Generated client accounts for UpdateTokenBondingV0.