Module mmoshforge::instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

BuyLaunchPass

Instruction.

BuyNativeV0

Instruction.

BuyV1

Instruction.

BuyWrappedSolV0

Instruction.

CloseTokenBondingV0

Instruction.

CreateCollection

Instruction.

CreateCurveV0

Instruction.

CreatePassToken

Instruction.

InitActivationToken

Instruction.

InitLaunchPass

Instruction.

InitMainState

Instruction.

InitPassToken

Instruction.

InitVault

Instruction.

InitializeSolStorageV0

Instruction.

InitializeTokenBondingV0

Instruction.

MintActivationToken

Instruction.

MintGenesisPass

Instruction.

MintGenesisProfile

Instruction.

MintGuestPass

Instruction.

MintPassByAt

Instruction.

MintProfileByAt

Instruction.

RedeemLaunchPass

Instruction.

SellNativeV0

Instruction.

SellV1

Instruction.

SellWrappedSolV0

Instruction.

SetCommonLut

Instruction.

StakeVault

Instruction.

TransferReservesNativeV0

Instruction.

TransferReservesV0

Instruction.

UnstakeVault

Instruction.

UpdateCurveV0

Instruction.

UpdateMainState

Instruction.

UpdateMainStateOwner

Instruction.

UpdateProfile

Instruction.

UpdateReserveAuthorityV0

Instruction.

UpdateTokenBondingV0

Instruction.